code with flag fic

+++ src/parsing/tmp/debug_map.c
@@ -6,7 +6,7 @@
 /*   By: sperron <sperron@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/18 17:00:04 by sperron           #+#    #+#             */
-/*   Updated: 2024/11/20 16:19:52 by sperron          ###   ########.fr       */
+/*   Updated: 2024/11/20 16:22:20 by sperron          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	debug_map(t_data *data)
 {
-	size_t	i;
+	int	i;
 
 	i = -1;
 	printf("=== Data Stats ===\n");
